Move include *.hpp to include/goodform/*hpp and add -Wall to compile flags
